add monitor

+++ src/main/java/transit/management/businesslayer/strategy/ConsumptionStrategy.java
@@ -1,5 +1,7 @@
 package transit.management.businesslayer.strategy;
 
+import java.math.BigDecimal;
+
 public interface ConsumptionStrategy {
-    double calculateConsumption(double miles);
+    BigDecimal calculateConsumption(BigDecimal miles);
 }
